fix: BIG CHANGE. modified event participation logic by adding Attendee Model. Also cleaned and simplified models for messages, events and car groups.

+++ routes/car-group.routes.js
@@ -3,6 +3,7 @@ const router = express.Router();
 
 const CarGroup = require("../models/CarGroup.model")
 const Event = require("../models/Event.model")
+const Attendee = require("../models/Attendee.model")
 
 const validateMongoIdFormat = require("../utils/validateMongoIdFormat")
 const validateRequiredFields = require("../utils/validateRequiredFields")
@@ -12,9 +13,9 @@ const validateDateFormat = require("../utils/validateDateFormat")
 router.post("/:eventId", async (req, res, next) => {
 
   const { eventId } = req.params
-  const { pickupCoordinates, pickupLocation, pickupTime, roomAvailable } = req.body
+  const { roomAvailable, pickupLocation, pickupTime, carColor, carBrand } = req.body
 
-  const areRequiredFieldsValid = validateRequiredFields(res, roomAvailable, pickupLocation, pickupTime)
+  const areRequiredFieldsValid = validateRequiredFields(res, roomAvailable, pickupLocation, pickupTime, carColor, carBrand)
   if (!areRequiredFieldsValid) return
 
   const isEventIdValid = validateMongoIdFormat(eventId, res, "Id de evento en formato incorrecto")
@@ -23,34 +24,46 @@ router.post("/:eventId", async (req, res, next) => {
   try {
 
     const foundEvent = await Event.findById(eventId)
-    if (!foundEvent.participants.includes(req.payload._id)) {
+
+    if (!foundEvent) {
+      res.status(400).json({errorMessage: "No hay eventos con este id"})
+      return
+    }
+
+    if (!foundEvent.hasCarOrganization) {
+      res.status(400).json({errorMessage: "Este evento no permite organización de coches"})
+      return
+    }
+
+    const foundAttendee = await Attendee.findOne({event: eventId, user: req.payload._id})
+    console.log(foundAttendee)
+    if (!foundAttendee) {
       res.status(400).json({errorMessage: "No puedes crear grupo de coche porque no perteneces a este evento"})
       return
     }
 
-    if (foundEvent.category !== "car-group") {
-      res.status(400).json({errorMessage: "Categoria de evento no permite crear grupos de coche"})
-      return
-    }
-
-    const foundCarGroup = await CarGroup.findOne({$and: [{event: eventId}, {owner: req.payload._id}]})
+    const userOwnsCarGroupInEvent = {event: eventId, owner: req.payload._id}
+    const userIsInCarGroupInEvent = {event: eventId, passengers: {$in: req.payload._id}}
+    const foundCarGroup = await CarGroup.findOne({$or: [userOwnsCarGroupInEvent, userIsInCarGroupInEvent]})
     if (foundCarGroup) {
-      res.status(400).json({errorMessage: "Grupo de coche ya creado por este usuario para este evento"})
+      res.status(400).json({errorMessage: "No puedes crear grupo de coche porque ya creaste o perteneces a un grupo de coche para este evento"})
       return
     }
-
-    //todo no puedes crear grupo de coche si estas en otro
-    
-    await CarGroup.create({ 
-      pickupCoordinates,
+  
+    const createdCarGroup = await CarGroup.create({ 
       pickupLocation,
       pickupTime,
       roomAvailable,
+      carColor,
+      carBrand,
       owner: req.payload._id,
       event: eventId
     })
 
-    //todo maybe receive createdCar, validate 500 with problems creating car
+    if (!createdCarGroup) {
+      res.status(500).json({errorMessage: "Problema de base de datos creando el grupo de coche"})
+      return
+    }
 
     res.sendStatus(202)
 
@@ -72,8 +85,9 @@ router.get("/list/:eventId", async (req, res, next) => {
     
     const carGroupsByEvent = await CarGroup
       .find({ event: eventId })
-      .select("pickupLocation pickupCoordinates pickupTime roomAvailable members owner")
+      .select("owner passengers roomAvailable pickupLocation pickupTime carColor carBrand status")
       .populate("owner", "username fullName icon iconColor")
+      
     res.status(200).json(carGroupsByEvent)
 
   } catch (error) {
@@ -95,25 +109,15 @@ router.get("/:carGroupId", async (req, res, next) => {
     const carGroupDetails = await CarGroup
       .findById(carGroupId)
       .populate("owner", "username fullName phoneCode phoneNumber icon iconColor")
-      .populate("members", "username fullName phoneCode phoneNumber icon iconColor")
-      .populate({
-        path: "messages",
-        model: "Message",
-        populate: {
-          path: "sender",
-          model: "User",
-          select: "username fullName icon iconColor"
-        }
-      })
-
-    //todo populate the event and send to FE to display to which event is it for
+      .populate("passengers", "username fullName phoneCode phoneNumber icon iconColor")
+      .populate("event", "title location date time")
     
     if (!carGroupDetails) {
       res.status(400).json({errorMessage: "Grupo de coche no existe por ese id"})
       return
     }
 
-    const isUserInCarGroup = carGroupDetails.members.some((member) => member._id == req.payload._id)
+    const isUserInCarGroup = carGroupDetails.passengers.some((passenger) => passenger._id == req.payload._id)
    
     if (carGroupDetails.owner._id != req.payload._id && !isUserInCarGroup) {
       res.status(400).json({errorMessage: "No perteneces a este grupo de coche como dueño o como miembro"})
@@ -132,15 +136,14 @@ router.get("/:carGroupId", async (req, res, next) => {
 router.put("/:carGroupId", async (req, res, next) => {
 
   const { carGroupId } = req.params
-  const { pickupCoordinates, pickupLocation, pickupTime, roomAvailable } = req.body
+  const { pickupLocation, pickupTime, roomAvailable, carColor, carBrand } = req.body
 
   const isCarGroupIdValid = validateMongoIdFormat(carGroupId, res, "Id de evento en formato incorrecto")
   if (!isCarGroupIdValid) return
 
-  const areRequiredFieldsValid = validateRequiredFields(res, pickupLocation, roomAvailable, pickupTime)
+  const areRequiredFieldsValid = validateRequiredFields(res, pickupLocation, roomAvailable, pickupTime, carColor, carBrand)
   if (!areRequiredFieldsValid) return
 
-  
   try {
     
     const foundCarGroup = await CarGroup.findById(carGroupId)
@@ -155,13 +158,19 @@ router.put("/:carGroupId", async (req, res, next) => {
       return
     }
     
-    //todo no puedes disminuir plazas disponibles si ya estan llenas
+    const newRoomAvailableWithPassangers = roomAvailable - foundCarGroup.passengers.length
+
+    if (newRoomAvailableWithPassangers < 0) {
+      res.status(401).json({errorMessage: "No puedes reducir la cantidad de plazas del coche porque ya estan llenas. contacta a un pasajero o elimina el grupo de coche"})
+      return
+    }
 
     await CarGroup.findByIdAndUpdate(carGroupId, { 
-      pickupCoordinates,
       pickupLocation,
       pickupTime,
       roomAvailable,
+      carColor,
+      carBrand,
     })
 
     res.sendStatus(202)
@@ -181,42 +190,59 @@ router.patch("/:carGroupId/join", async (req, res, next) => {
 
   try {
 
-    const carGroup = await CarGroup.findById(carGroupId)
+    const carGroup = await CarGroup.findById(carGroupId).populate("event")
 
     if (!carGroup) {
       res.status(400).json({ errorMessage: "No hay grupos de coche con ese id" })
       return
     }
 
-    if (carGroup.members.length >= carGroup.roomAvailable ) {
+    if (carGroup.isCancelled) {
+      res.status(400).json({ errorMessage: "Este Grupo de coche ha sido cancelado" })
+      return
+    }
+
+    if (carGroup.passengers.length >= carGroup.roomAvailable ) {
       res.status(400).json({ errorMessage: "No hay espacio disponible en este grupo de coche" })
       return
     }
 
     if (carGroup.owner._id == req.payload._id) {
-      res.status(400).json({ errorMessage: "No te puedes unir como miembro al grupo de coche que has creado" })
+      res.status(400).json({ errorMessage: "No te puedes unir como pasajero al grupo de coche que has creado" })
       return
     }
 
-    const event = await Event.findById(carGroup.event)
+    if (carGroup.passengers.includes(req.payload._id)) {
+      res.status(400).json({ errorMessage: "Ya te has unido a este grupo de coche" })
+      return
+    }
 
-    if (!event || event.isCancelled) {
+    const { event } = carGroup
+
+    if (!event || event.status === "cancelled") {
       res.status(400).json({ errorMessage: "No te puedes unir al grupo de coche porque el evento de ese grupo de coche no existe o ha sido cancelado" })
       return
     }
 
-    if (!event.participants.includes(req.payload._id)) {
+    const attendee = await Attendee.findOne({ user: req.payload._id, event: event._id })
+    if (!attendee) {
       res.status(400).json({ errorMessage: "No te puedes unir al grupo de coche porque no te has unido al evento" })
       return
     }
 
-    const carGroupAlreadyJoined = await CarGroup.findOne({$and: [{event: event._id}, {members: {$in: req.payload._id}}]})
-    if (carGroupAlreadyJoined) {
-      res.status(400).json({ errorMessage: "No te puedes unir al grupo de coche porque ya estas en un grupo de coche para el mismo evento." })
+    const existingCarGroup = await CarGroup.findOne({
+      event: event._id,
+      $or: [
+        { owner: req.payload._id }, // User is the owner
+        { passengers: { $in: [req.payload._id] } } // User is a passenger
+      ]
+    });
+    if (existingCarGroup) {
+      res.status(400).json({ errorMessage: "No te puedes unir al grupo de coche porque ya estas en un grupo de coche para el mismo evento o tienes uno propio creado" })
       return
     }
 
-    const updatedCarGroup = await CarGroup.findByIdAndUpdate(carGroupId, { $addToSet: { members: req.payload._id } })
+    const updatedCarGroup = await CarGroup.findByIdAndUpdate(carGroupId, { $addToSet: { passengers: req.payload._id } })
 
     if (!updatedCarGroup) {
       res.status(500).json({ errorMessage: "Hubo un problema al agregar el usuario al grupo de coche" })
@@ -241,10 +267,10 @@ router.patch("/:carGroupId/leave", async (req, res, next) => {
 
   try {
 
-    const updatedCarGroup = await CarGroup.findByIdAndUpdate(carGroupId, { $pull: { members: req.payload._id } })
+    const updatedCarGroup = await CarGroup.findByIdAndUpdate(carGroupId, { $pull: { passengers: req.payload._id } })
 
     if (!updatedCarGroup) {
-      res.status(400).json({ errorMessage: "No hay eventos con ese id o hubo un problema al remover al usuario del grupo de coche" })
+      res.status(400).json({ errorMessage: "No hay grupos de coche con ese id o hubo un problema al remover al usuario del grupo de coche" })
       return
     }
 
@@ -261,8 +287,6 @@ router.patch("/:carGroupId/leave", async (req, res, next) => {
 // DELETE "/api/car-group/:carGroupId" - Attemps to delete a car-group
 router.delete("/:carGroupId", async (req, res, next) => {
 
-  //*IMPORTANT: car groups will not have a status to preserve data, instead, notifications will handle if a car-group a user was in was deleted
-
   const { carGroupId } = req.params
 
   const isCarGroupIdValid = validateMongoIdFormat(carGroupId, res, "Id de evento en formato incorrecto")
@@ -270,7 +294,7 @@ router.delete("/:carGroupId", async (req, res, next) => {
 
   try {
     
-    const foundCarGroup = await CarGroup.findById(carGroupId)
+    const foundCarGroup = await CarGroup.findById(carGroupId).populate("event", "status")
 
     if (!foundCarGroup) {
       res.status(400).json({ errorMessage: "No hay grupos de coche con ese id" })
@@ -282,10 +306,20 @@ router.delete("/:carGroupId", async (req, res, next) => {
       return
     }
 
-    await CarGroup.findByIdAndDelete(carGroupId)
+    if (foundCarGroup.event.status === "cancelled") {
+      res.status(400).json({ errorMessage: "Este evento ha sido cancelado, no es necesario borrar grupo de coche" })
+      return
+    }
 
-    //todo borrar todos los mensajes de ese grupo de coche
+    if (foundCarGroup.event.status === "closed") {
+      res.status(400).json({ errorMessage: "Este evento está cerrado, no puedes borrar el grupo de coche. Contacta a un Admin o el creador del evento para poder borrarlo" })
+      return
+    }
+
+    //* open event only
+    await CarGroup.findByIdAndDelete(carGroupId)
     //todo notificacion a usuarios si se borra el grupo de coche aqui
+    //todo borrar todos los mensajes de ese grupo de coche
 
     res.sendStatus(202)
